rename sm_messages to messages
