+++ packages/chain/statemanager/gpa/messages/block_message.go
@@ -1,4 +1,4 @@
-package sm_messages
+package messages
 
 import (
 	bcs "github.com/iotaledger/bcs-go"
