+++ packages/chain/statemanager/gpa/messages/block_message_test.go
@@ -1,4 +1,4 @@
-package sm_messages
+package messages
 
 import (
 	"testing"
